Attempting to manually change gitignore didn't work
It's still recognizing the Unity Lockfile as a change on mine, so I have
to close Unity to push things.
A few other things are also listed in the changes that really don't need
to be there.
Do any of you guys know how to fix this?

(I just changed the scale of the dummy button I made earlier to test
this)